msvc capatable.

+++ src/ethertia/util/Collections.h
@@ -52,7 +52,7 @@ public:
     }
 
     template<typename T>
-    static T* range(T* arr, int n, T from = 0) {
+    static T* range(T* arr, size_t n, T from = 0) {
         for (int i = 0;i < n;++i) {
             arr[i] = from + i;
         }
